feat: integrate places with tags
- Add new fields to Places model
- Add default tags
- Fix PlaceTags model to support BIGINT place_id
- Update places service to include tags in responses

+++ models/placetags.js
@@ -22,7 +22,7 @@ module.exports = (sequelize, DataTypes) => {
   }
   PlaceTags.init(
     {
-      place_id: DataTypes.INTEGER,
+      place_id: DataTypes.BIGINT,
       tag_id: DataTypes.INTEGER,
     },
     {
